refactor: DTO 이름 및 패키지 이름 변경
