+++ src/main/java/wooteco/subway/dto/response/StationResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.response;
 
 import wooteco.subway.domain.Station;
