+++ src/test/java/wooteco/subway/acceptance/LineAcceptanceTest.java
@@ -21,7 +21,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.dto.StationResponse;
+import wooteco.subway.dto.response.StationResponse;
 import wooteco.subway.dto.response.LineResponse;
 
 @Nested
